refactor: split session access controls and tests
