+++ ui/src/lib/sessions/index.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import {
   GatewayRequestError,
   type GatewayBrowserClient,
@@ -7,7 +7,6 @@ import {
   type GatewayHelloOk,
 } from "../../api/gateway.ts";
 import type { SessionsListResult } from "../../api/types.ts";
-import { createStorageMock } from "../../test-helpers/storage.ts";
 import { waitForFast } from "../../test-helpers/wait-for.ts";
 import { createSessionCapability, reconcileSessionRunTerminal } from "./index.ts";
 
@@ -31,11 +30,7 @@ function deferred<T>() {
   return { promise, reject, resolve };
 }
 
-function createGatewayHarness(
-  client: GatewayBrowserClient,
-  featureMethods?: string[],
-  scopes?: string[],
-) {
+function createGatewayHarness(client: GatewayBrowserClient, featureMethods?: string[]) {
   let snapshot: {
     client: GatewayBrowserClient | null;
     phase: "connected" | "reconnecting";
@@ -50,10 +45,7 @@ function createGatewayHarness(
     hello:
       featureMethods === undefined
         ? null
-        : ({
-            ...(scopes ? { auth: { role: "operator", scopes } } : {}),
-            features: { methods: featureMethods },
-          } as GatewayHelloOk),
+        : ({ features: { methods: featureMethods } } as GatewayHelloOk),
   };
   const listeners = new Set<(next: typeof snapshot) => void>();
   const eventListeners = new Set<(event: GatewayEventFrame) => void>();
@@ -88,10 +80,6 @@ function createGatewayHarness(
     },
   };
 }
-
-afterEach(() => {
-  vi.unstubAllGlobals();
-});
 
 function sessionChangedEvent(key: string): GatewayEventFrame {
   return {
@@ -179,61 +167,6 @@ describe("createSessionCapability", () => {
     await sessions.groupsLoad();
 
     expect(request).toHaveBeenCalledOnce();
-    sessions.dispose();
-  });
-
-  it("does not migrate legacy browser groups without operator.write", async () => {
-    vi.stubGlobal("localStorage", createStorageMock());
-    localStorage.setItem("openclaw:sessions:custom-groups", JSON.stringify(["Research"]));
-    const request = vi.fn(async (method: string) => {
-      if (method === "sessions.groups.list") {
-        return { groups: [] };
-      }
-      throw new Error(`Unexpected request: ${method}`);
-    });
-    const client = { request } as unknown as GatewayBrowserClient;
-    const { gateway } = createGatewayHarness(
-      client,
-      ["sessions.groups.list", "sessions.groups.put"],
-      ["operator.read"],
-    );
-    const sessions = createSessionCapability(gateway);
-
-    await sessions.groupsLoad();
-
-    expect(request).toHaveBeenCalledOnce();
-    expect(request).toHaveBeenCalledWith("sessions.groups.list", {});
-    expect(localStorage.getItem("openclaw:sessions:custom-groups")).toBe(
-      JSON.stringify(["Research"]),
-    );
-    sessions.dispose();
-  });
-
-  it("migrates legacy browser groups with operator.write", async () => {
-    vi.stubGlobal("localStorage", createStorageMock());
-    localStorage.setItem("openclaw:sessions:custom-groups", JSON.stringify(["Research"]));
-    const request = vi.fn(async (method: string) => {
-      if (method === "sessions.groups.list") {
-        return { groups: [] };
-      }
-      if (method === "sessions.groups.put") {
-        return { groups: [{ name: "Research" }] };
-      }
-      throw new Error(`Unexpected request: ${method}`);
-    });
-    const client = { request } as unknown as GatewayBrowserClient;
-    const { gateway } = createGatewayHarness(
-      client,
-      ["sessions.groups.list", "sessions.groups.put"],
-      ["operator.write"],
-    );
-    const sessions = createSessionCapability(gateway);
-
-    await sessions.groupsLoad();
-
-    expect(request).toHaveBeenCalledWith("sessions.groups.put", { names: ["Research"] });
-    expect(sessions.state.groups).toEqual(["Research"]);
-    expect(localStorage.getItem("openclaw:sessions:custom-groups")).toBeNull();
     sessions.dispose();
   });
